requirements.txt
adal==1.2.7
asgiref==3.3.4
Beaker==1.7.0
beautifulsoup4==4.9.3
bs4==0.0.1
bump2version==1.0.1
bumpversion==0.6.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2.3
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-request==1.5.6
fhirclient==3.2.0
Flask==0.10.1
gunicorn==20.1.0
idna==2.10
isodate==0.5.1
itsdangerous==0.24
Jinja2==2.11.3
jsontree==0.5.1
lxml==4.6.3
MarkupSafe==0.23
oauthlib==3.1.1
pafy==0.5.5
path==15.1.2
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
pytube==10.8.4
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tk==0.1.0
urllib3==1.26.4
Werkzeug==0.14.1
whitenoise==3.3.1
xmljson==0.2.1
xmltodict==0.12.0
youtube-dl==2021.6.6
